requirements.txt
beautifulsoup4==4.13.5
ebooklib==0.19
fitz==0.0.1.dev2
Flask==3.1.2
flask_cors==6.0.1
langchain==0.3.27
langchain_mistralai==0.2.11
numpy==2.3.3
odfpy==1.4.1
openpyxl==3.1.5
pdfplumber==0.11.7
PyJWT==2.10.1
python-dotenv==1.1.1
python_docx==1.2.0
python_pptx==1.0.2
qdrant_client==1.15.1
rtfparse==0.9.5
sentence_transformers==5.1.0
Werkzeug==3.1.3

# Core functionality dependencies
requests==2.32.3
rank_bm25==0.2.2
spacy==3.8.2
